+++ example/double_stdout.conf
