+++ example/stdout_and_file.conf
@@ -1,0 +1,25 @@
+<source>
+  @type dummy
+  tag "test.*"
+</source>
+<match **>
+  @type copy_ex
+  <store>
+    @type stdout
+  </store>
+  <store>
+    @type file
+    append true
+    path "output.log"
+    time_slice_format test-%Y-%m-%d
+    <format>
+      @type json
+      localtime false
+    </format>
+    <buffer>
+      @type memory
+      flush_mode immediate
+      flush_at_shutdown true
+    </buffer>
+  </store>
+</match>
